client/requirements.txt
requests
firebird-driver
psycopg2-binary
configparser